# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IPL Auction Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Season {
  year: number;
  name: string;
  totalTeams: number;
  totalPlayers: number;
}

export interface Franchise {
  id: number;
  name: string;
  shortName: string;
  city: string;
  primaryColor: string;
  secondaryColor: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface Player {
  id: number;
  name: string;
  role: string;
  nationality: string;
  isOverseas: boolean;
  /** @nullable */
  age?: number | null;
  basePricecrore: number;
  /** @nullable */
  soldPricecrore?: number | null;
  isCapped: boolean;
  /** @nullable */
  franchiseId?: number | null;
  /** @nullable */
  franchiseName?: string | null;
  /** @nullable */
  battingStyle?: string | null;
  /** @nullable */
  bowlingStyle?: string | null;
}

export type RoomInputAuctionType = typeof RoomInputAuctionType[keyof typeof RoomInputAuctionType];


export const RoomInputAuctionType = {
  mega: 'mega',
  mini: 'mini',
} as const;

export type RoomInputAuctionSpeed = typeof RoomInputAuctionSpeed[keyof typeof RoomInputAuctionSpeed];


export const RoomInputAuctionSpeed = {
  fast: 'fast',
  normal: 'normal',
  slow: 'slow',
} as const;

export interface RoomInput {
  name: string;
  seasonYear: number;
  auctionType: RoomInputAuctionType;
  budgetCrore: number;
  maxSquadSize: number;
  maxOverseas: number;
  maxOwnersPerTeam: number;
  auctionSpeed: RoomInputAuctionSpeed;
  rtmEnabled?: boolean;
  maxRetentions?: number;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  lobby: 'lobby',
  preparation: 'preparation',
  auction: 'auction',
  completed: 'completed',
} as const;

export interface Room {
  id: number;
  code: string;
  name: string;
  hostUserId: string;
  seasonYear: number;
  auctionType: string;
  status: RoomStatus;
  budgetCrore: number;
  maxSquadSize: number;
  maxOverseas: number;
  maxOwnersPerTeam: number;
  auctionSpeed: string;
  currentPlayerIndex: number;
  rtmEnabled?: boolean;
  maxRetentions?: number;
  /** @nullable */
  rtmPendingTeamId?: number | null;
  /** @nullable */
  rtmBidAmountCrore?: number | null;
  /** @nullable */
  rtmBidderTeamId?: number | null;
  createdAt: string;
}

export interface JoinRoomInput {
  userId: string;
  displayName: string;
}

export interface RoomMember {
  id: number;
  roomId: number;
  userId: string;
  displayName: string;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  isOnline: boolean;
  joinedAt: string;
}

export interface TeamBudgetSummary {
  teamId: number;
  franchiseName: string;
  shortName: string;
  primaryColor: string;
  budgetRemaining: number;
  budgetSpent: number;
  squadSize: number;
  overseasCount: number;
}

export interface RoomSummary {
  room: Room;
  teamCount: number;
  memberCount: number;
  playersAuctioned: number;
  playersRemaining: number;
  totalMoneySpent: number;
  teams: TeamBudgetSummary[];
}

export interface TeamOwner {
  userId: string;
  displayName: string;
  isOnline: boolean;
}

export interface Team {
  id: number;
  roomId: number;
  franchiseId: number;
  franchiseName: string;
  shortName: string;
  primaryColor: string;
  secondaryColor: string;
  budgetRemainingCrore: number;
  budgetSpentCrore: number;
  retentionComplete: boolean;
  owners: TeamOwner[];
}

export interface RetentionInput {
  playerIds: number[];
  retentionPricesCrore: number[];
}

export interface SquadPlayer {
  id: number;
  name: string;
  role: string;
  nationality: string;
  isOverseas: boolean;
  priceCrore: number;
  isRetained: boolean;
  /** @nullable */
  battingStyle?: string | null;
  /** @nullable */
  bowlingStyle?: string | null;
}

export interface TeamSquad {
  team: Team;
  retainedPlayers: SquadPlayer[];
  auctionedPlayers: SquadPlayer[];
  batters: number;
  bowlers: number;
  allRounders: number;
  wicketKeepers: number;
  overseas: number;
  indian: number;
  /** @nullable */
  averageAge?: number | null;
}

export type AuctionPoolPlayerStatus = typeof AuctionPoolPlayerStatus[keyof typeof AuctionPoolPlayerStatus];


export const AuctionPoolPlayerStatus = {
  available: 'available',
  sold: 'sold',
  unsold: 'unsold',
  retained: 'retained',
} as const;

export interface AuctionPoolPlayer {
  id: number;
  playerId: number;
  name: string;
  role: string;
  nationality: string;
  isOverseas: boolean;
  basePriceCrore: number;
  status: AuctionPoolPlayerStatus;
  /** @nullable */
  soldToCrore?: number | null;
  /** @nullable */
  soldToTeamId?: number | null;
  /** @nullable */
  soldToTeamName?: string | null;
  /** @nullable */
  previousTeamName?: string | null;
}

export type AuctionStateStatus = typeof AuctionStateStatus[keyof typeof AuctionStateStatus];


export const AuctionStateStatus = {
  idle: 'idle',
  bidding: 'bidding',
  rtm: 'rtm',
  sold: 'sold',
  unsold: 'unsold',
  completed: 'completed',
} as const;

export interface AuctionState {
  currentPlayer?: AuctionPoolPlayer;
  currentBidCrore: number;
  /** @nullable */
  currentBidderTeamId?: number | null;
  /** @nullable */
  currentBidderTeamName?: string | null;
  timerSeconds: number;
  status: AuctionStateStatus;
  /** @nullable */
  rtmPendingTeamId?: number | null;
  /** @nullable */
  rtmPendingTeamName?: string | null;
  totalPlayersAuctioned: number;
  totalPlayersRemaining: number;
}

export interface BidInput {
  teamId: number;
  bidAmountCrore: number;
}

export interface Bid {
  id: number;
  teamId: number;
  teamName: string;
  bidAmountCrore: number;
  placedAt: string;
}

export interface Message {
  id: number;
  roomId: number;
  /** @nullable */
  teamId?: number | null;
  userId: string;
  displayName: string;
  content: string;
  isSystem: boolean;
  createdAt: string;
}

export interface MessageInput {
  userId: string;
  displayName: string;
  content: string;
  /** @nullable */
  teamId?: number | null;
}

export type AiReportStatus = typeof AiReportStatus[keyof typeof AiReportStatus];


export const AiReportStatus = {
  pending: 'pending',
  generating: 'generating',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface TeamRanking {
  rank: number;
  teamId: number;
  franchiseName: string;
  shortName: string;
  primaryColor: string;
  overallScore: number;
  topOrderScore: number;
  middleOrderScore: number;
  finishersScore: number;
  spinAttackScore: number;
  paceAttackScore: number;
  deathBowlingScore: number;
  powerpayBowlingScore: number;
  benchStrengthScore: number;
  captaincyScore: number;
  balanceScore: number;
  squadDepthScore: number;
  strengths: string[];
  weaknesses: string[];
  reasoning: string;
  budgetUtilization: number;
}

export interface AiReport {
  id: number;
  roomId: number;
  status: AiReportStatus;
  /** @nullable */
  generatedAt?: string | null;
  rankings: TeamRanking[];
  /** @nullable */
  predictedChampion?: string | null;
  /** @nullable */
  overallAnalysis?: string | null;
}

export type ListSeasonPlayersParams = {
/**
 * @nullable
 */
teamId?: number | null;
/**
 * @nullable
 */
role?: string | null;
/**
 * @nullable
 */
nationality?: string | null;
};

export type ListAuctionPoolParams = {
/**
 * @nullable
 */
role?: string | null;
/**
 * @nullable
 */
nationality?: string | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
search?: string | null;
};

export type MakeRtmDecisionBody = {
  useRtm: boolean;
};

export type ListRoomMessagesParams = {
/**
 * @nullable
 */
teamId?: number | null;
};