# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IPL Auction Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create an auction room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{code}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
  /rooms/{code}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a room
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomInput"
      responses:
        "200":
          description: Joined room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomMember"
  /rooms/{code}/members:
    get:
      operationId: getRoomMembers
      tags: [rooms]
      summary: Get room members
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomMember"
  /rooms/{code}/start:
    post:
      operationId: startAuction
      tags: [rooms]
      summary: Start the auction
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Auction started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{code}/summary:
    get:
      operationId: getRoomSummary
      tags: [rooms]
      summary: Get room summary with budget overview
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSummary"
...
components:
  schemas:
...
    RoomInput:
      type: object
      properties:
        name:
          type: string
        seasonYear:
          type: integer
        auctionType:
          type: string
          enum: [mega, mini]
        budgetCrore:
          type: number
        maxSquadSize:
          type: integer
        maxOverseas:
          type: integer
        maxOwnersPerTeam:
          type: integer
        auctionSpeed:
          type: string
          enum: [fast, normal, slow]
        rtmEnabled:
          type: boolean
          default: true
        maxRetentions:
          type: integer
          default: 6
      required: [name, seasonYear, auctionType, budgetCrore, maxSquadSize, maxOverseas, maxOwnersPerTeam, auctionSpeed]
    Room:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        hostUserId:
          type: string
        seasonYear:
          type: integer
        auctionType:
          type: string
        status:
          type: string
          enum: [lobby, preparation, auction, completed]
        budgetCrore:
          type: number
        maxSquadSize:
          type: integer
        maxOverseas:
          type: integer
        maxOwnersPerTeam:
          type: integer
        auctionSpeed:
          type: string
        currentPlayerIndex:
          type: integer
        rtmEnabled:
          type: boolean
        maxRetentions:
          type: integer
        rtmPendingTeamId:
          type: ["integer", "null"]
        rtmBidAmountCrore:
          type: ["number", "null"]
        rtmBidderTeamId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, code, name, hostUserId, seasonYear, auctionType, status, budgetCrore, maxSquadSize, maxOverseas, maxOwnersPerTeam, auctionSpeed, currentPlayerIndex, createdAt]
    JoinRoomInput:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
      required: [userId, displayName]
    RoomMember:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        isOnline:
          type: boolean
        joinedAt:
          type: string
      required: [id, roomId, userId, displayName, isOnline, joinedAt]
    RoomSummary:
      type: object
      properties:
        room:
          $ref: "#/components/schemas/Room"
        teamCount:
          type: integer
        memberCount:
          type: integer
        playersAuctioned:
          type: integer
        playersRemaining:
          type: integer
        totalMoneySpent:
          type: number
        teams:
          type: array
          items:
            $ref: "#/components/schemas/TeamBudgetSummary"
      required: [room, teamCount, memberCount, playersAuctioned, playersRemaining, totalMoneySpent, teams]
    TeamBudgetSummary:
      type: object
      properties:
        teamId:
          type: integer
        franchiseName:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        budgetRemaining:
          type: number
        budgetSpent:
          type: number
        squadSize:
          type: integer
        overseasCount:
          type: integer
      required: [teamId, franchiseName, shortName, primaryColor, budgetRemaining, budgetSpent, squadSize, overseasCount]
...