# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IPL Auction Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  code: string;
  name: string;
  hostUserId: string;
  seasonYear: number;
  auctionType: string;
  status: RoomStatus;
  budgetCrore: number;
  maxSquadSize: number;
  maxOverseas: number;
  maxOwnersPerTeam: number;
  auctionSpeed: string;
  currentPlayerIndex: number;
  rtmEnabled?: boolean;
  maxRetentions?: number;
  /** @nullable */
  rtmPendingTeamId?: number | null;
  /** @nullable */
  rtmBidAmountCrore?: number | null;
  /** @nullable */
  rtmBidderTeamId?: number | null;
  createdAt: string;
}