# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IPL Auction Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomInputAuctionSpeed } from './roomInputAuctionSpeed';
import type { RoomInputAuctionType } from './roomInputAuctionType';

export interface RoomInput {
  name: string;
  seasonYear: number;
  auctionType: RoomInputAuctionType;
  budgetCrore: number;
  maxSquadSize: number;
  maxOverseas: number;
  maxOwnersPerTeam: number;
  auctionSpeed: RoomInputAuctionSpeed;
  rtmEnabled?: boolean;
  maxRetentions?: number;
}